+++ Equation.cs
@@ -4,8 +4,8 @@ namespace WpfApp_Calc
 {
     public class Equation
     {
-        public List<double> Numbers { get; set; } = new();
+        public List<decimal> Numbers { get; set; } = new();
         public List<string> Symbols { get; set; } = new();
-        public double Result { get; set; }
+        public decimal Result { get; set; }
     }
 }
